% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anotworking.R
\name{Anotworking}
\alias{Anotworking}
\title{Problem A}
\usage{
Anotworking()
}
\description{
\loadmathjax
Example of equation in \code{Details} not converting properly to the Rd file when using \code{devtools::document()}
}
\details{
\mjdeqn{\hat{p}\emph{c = \frac{1}{N} \sum}{i=1}^N B_{i,c}}{ascii}
}
