% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aworking.R
\name{Aworking}
\alias{Aworking}
\title{Modified A so there's no longer a problem}
\usage{
Aworking()
}
\description{
\loadmathjax
Example of where the equation in \code{Anotworking} does work: the p without a hat
}
\details{
\mjdeqn{p_{c} = \frac{1}{N} \sum_{i=1}^N B_{i,c}}{ascii}
}
