% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bnotworking.R
\name{Bnotworking}
\alias{Bnotworking}
\title{Problem B}
\usage{
Bnotworking()
}
\description{
\loadmathjax
Example of equation in \code{Details} not converting properly to the Rd file when using \code{devtools::document()}
}
\details{
\mjdeqn{w\\emph{outlier}{i,c}B_{i,c}=1-\frac{d_{i,c}}{\lbrace {d_{j,c}} \rbrace}\lbrace {d_{j,c}}\rbrace}{ascii}
}
