% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bworking.R
\name{Bworking}
\alias{Bworking}
\title{Modified B so there's no longer a problem}
\usage{
Bworking()
}
\description{
\loadmathjax
Example of where the equation in \code{Bnotworking} does work: replacing underscore with space
}
\details{
\mjdeqn{w\text{ }outlier_{i,c}B_{i,c}=1-\frac{d_{i,c}}{\lbrace {d_{j,c}} \rbrace}\lbrace {d_{j,c}}\rbrace}{ascii}
}
